Se agrega end point para agregar género

+++ UniversidadApi/Models/Genero.cs
@@ -1,11 +1,11 @@
 ﻿namespace UniversidadApi.Models
 {
-    public class Genero : BaseEntity<int>
+    public class Genero : BaseEntity<short>
     {
         public string Nombre { get; set; } = string.Empty;
         public string Sigla { get; set; } = string.Empty;
 
-        public Genero(int id, string nombre, string sigla, bool activo)
+        public Genero(short id, string nombre, string sigla, bool activo)
         {
             Id = id;
             Nombre = nombre;
