+++ UniversidadApi/Services/GeneroService/IGeneroService.cs
@@ -5,7 +5,7 @@ namespace UniversidadApi.Services.GeneroService
     public interface IGeneroService
     {
         Task<List<Genero>> GetAllGeneros();
-        Task<Genero?> GetGenero(int id);
+        Task<Genero?> GetGenero(short id);
         Task<Genero?> AddGenero (Genero entity);
         Task<Genero?> UpdateGenero(Genero entity);
     }
